Some minor tweaks and changing the health to float values

+++ Assets/Scripts/Enemies and Projectiles/DummyEnemy.cs
@@ -49,7 +49,7 @@ namespace AssemblyCSharp
 
           }
 
-          public int currentHp()
+          public float currentHp()
           {
                return health.currentHealth;
           }
