+++ Assets/Scripts/Enemies and Projectiles/Flamie.cs
@@ -143,7 +143,7 @@ namespace AssemblyCSharp
                return isAgro;
           }
 
-          public int currentHp()
+          public float currentHp()
           {
                return health.currentHealth;
           }
